requirements.txt
Brotli==1.0.9
click==7.1.2
colored==1.4.2
css==0.1
cx-Oracle==8.1.0
cycler==0.10.0
Cython==0.29.23
dash==1.20.0
dash-bootstrap-components==0.12.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
database-connection==0.1.1
database-connector==2.5.0
dill==0.3.3
Flask==1.1.2
Flask-Compress==1.9.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
Flask-SQLAlchemy-Session==1.1
future==0.18.2
html==1.13
itsdangerous==1.1.0
javascript==2015.32.59
Jinja2==2.11.3
kiwisolver==1.3.1
logger==1.4
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.20.2
packaging==20.9
pandas==1.2.4
php==1.2.1
plotly==4.14.3
POT==0.7.0
protobuf==3.15.6
psycopg2==2.8.6
psycopg2-binary==2.8.6
pymongo==3.11.3
PyMySQL==1.0.2
PyMysqlDB==0.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
retrying==1.3.3
route==0.1.4
scipy==1.6.2
Send2Trash==1.5.0
six==1.15.0
sql-alchemy-db==0.0.4
SQLAlchemy==1.3.10
sqlalchemy-redshift==0.8.2
sqlite-connector==0.1.0
sqlite-db-tools==0.1
termcolor==1.1.0
Werkzeug==1.0.1
